noop separate deploy.go and deploy_rp.go
