+++ pkg/deploy/deploy.go
@@ -5,19 +5,12 @@ package deploy
 
 import (
 	"context"
-	"encoding/base64"
-	"encoding/json"
 	"reflect"
 	"strings"
 
-	mgmtdns "github.com/Azure/azure-sdk-for-go/services/dns/mgmt/2018-05-01/dns"
-	mgmtfeatures "github.com/Azure/azure-sdk-for-go/services/resources/mgmt/2019-07-01/features"
-	"github.com/Azure/go-autorest/autorest/azure"
 	"github.com/Azure/go-autorest/autorest/azure/auth"
-	"github.com/Azure/go-autorest/autorest/to"
 	"github.com/sirupsen/logrus"
 
-	"github.com/Azure/ARO-RP/pkg/deploy/generator"
 	"github.com/Azure/ARO-RP/pkg/env"
 	"github.com/Azure/ARO-RP/pkg/util/arm"
 	"github.com/Azure/ARO-RP/pkg/util/azureclient/mgmt/compute"
@@ -103,143 +96,6 @@ func New(ctx context.Context, log *logrus.Entry, env env.Core, config *RPConfig,
 		config:  config,
 		version: version,
 	}, nil
-}
-
-func (d *deployer) Deploy(ctx context.Context) error {
-	encryptionAtHostSupported, err := d.encryptionAtHostSupported(ctx)
-	if err != nil {
-		return err
-	}
-	if !encryptionAtHostSupported {
-		d.log.Warn("encryption at host not supported")
-	}
-
-	rpMSI, err := d.userassignedidentities.Get(ctx, d.config.ResourceGroupName, "aro-rp-"+d.config.Location)
-	if err != nil {
-		return err
-	}
-
-	deploymentName := "rp-production-" + d.version
-
-	b, err := Asset(generator.FileRPProduction)
-	if err != nil {
-		return err
-	}
-
-	var template map[string]interface{}
-	err = json.Unmarshal(b, &template)
-	if err != nil {
-		return err
-	}
-
-	parameters := d.getParameters(template["parameters"].(map[string]interface{}))
-	parameters.Parameters["adminApiCaBundle"] = &arm.ParametersParameter{
-		Value: base64.StdEncoding.EncodeToString([]byte(*d.config.Configuration.AdminAPICABundle)),
-	}
-	parameters.Parameters["encryptionAtHost"] = &arm.ParametersParameter{
-		Value: encryptionAtHostSupported,
-	}
-	parameters.Parameters["extraCosmosDBIPs"] = &arm.ParametersParameter{
-		Value: strings.Join(d.config.Configuration.ExtraCosmosDBIPs, ","),
-	}
-	parameters.Parameters["rpImage"] = &arm.ParametersParameter{
-		Value: *d.config.Configuration.RPImagePrefix + ":" + d.version,
-	}
-	parameters.Parameters["rpServicePrincipalId"] = &arm.ParametersParameter{
-		Value: rpMSI.PrincipalID.String(),
-	}
-	parameters.Parameters["vmssName"] = &arm.ParametersParameter{
-		Value: d.version,
-	}
-
-	for i := 0; i < 2; i++ {
-		d.log.Printf("deploying %s", deploymentName)
-		err = d.deployments.CreateOrUpdateAndWait(ctx, d.config.ResourceGroupName, deploymentName, mgmtfeatures.Deployment{
-			Properties: &mgmtfeatures.DeploymentProperties{
-				Template:   template,
-				Mode:       mgmtfeatures.Incremental,
-				Parameters: parameters.Parameters,
-			},
-		})
-		if serviceErr, ok := err.(*azure.ServiceError); ok &&
-			serviceErr.Code == "DeploymentFailed" &&
-			i < 1 {
-			// on new RP deployments, we get a spurious DeploymentFailed error
-			// from the Microsoft.Insights/metricAlerts resources indicating
-			// that rp-lb can't be found, even though it exists and the
-			// resources correctly have a dependsOn stanza referring to it.
-			// Retry once.
-			d.log.Print(err)
-			continue
-		}
-		if err != nil {
-			return err
-		}
-
-		break
-	}
-
-	return d.configureDNS(ctx)
-}
-
-func (d *deployer) configureDNS(ctx context.Context) error {
-	rpPip, err := d.publicipaddresses.Get(ctx, d.config.ResourceGroupName, "rp-pip", "")
-	if err != nil {
-		return err
-	}
-
-	portalPip, err := d.publicipaddresses.Get(ctx, d.config.ResourceGroupName, "portal-pip", "")
-	if err != nil {
-		return err
-	}
-
-	zone, err := d.zones.Get(ctx, d.config.ResourceGroupName, d.config.Location+"."+*d.config.Configuration.ClusterParentDomainName)
-	if err != nil {
-		return err
-	}
-
-	_, err = d.globalrecordsets.CreateOrUpdate(ctx, *d.config.Configuration.GlobalResourceGroupName, *d.config.Configuration.RPParentDomainName, "rp."+d.config.Location, mgmtdns.A, mgmtdns.RecordSet{
-		RecordSetProperties: &mgmtdns.RecordSetProperties{
-			TTL: to.Int64Ptr(3600),
-			ARecords: &[]mgmtdns.ARecord{
-				{
-					Ipv4Address: rpPip.IPAddress,
-				},
-			},
-		},
-	}, "", "")
-	if err != nil {
-		return err
-	}
-
-	_, err = d.globalrecordsets.CreateOrUpdate(ctx, *d.config.Configuration.GlobalResourceGroupName, *d.config.Configuration.RPParentDomainName, d.config.Location+".admin", mgmtdns.A, mgmtdns.RecordSet{
-		RecordSetProperties: &mgmtdns.RecordSetProperties{
-			TTL: to.Int64Ptr(3600),
-			ARecords: &[]mgmtdns.ARecord{
-				{
-					Ipv4Address: portalPip.IPAddress,
-				},
-			},
-		},
-	}, "", "")
-	if err != nil {
-		return err
-	}
-
-	nsRecords := make([]mgmtdns.NsRecord, 0, len(*zone.NameServers))
-	for i := range *zone.NameServers {
-		nsRecords = append(nsRecords, mgmtdns.NsRecord{
-			Nsdname: &(*zone.NameServers)[i],
-		})
-	}
-
-	_, err = d.globalrecordsets.CreateOrUpdate(ctx, *d.config.Configuration.GlobalResourceGroupName, *d.config.Configuration.ClusterParentDomainName, d.config.Location, mgmtdns.NS, mgmtdns.RecordSet{
-		RecordSetProperties: &mgmtdns.RecordSetProperties{
-			TTL:       to.Int64Ptr(3600),
-			NsRecords: &nsRecords,
-		},
-	}, "", "")
-	return err
 }
 
 // getParameters returns an *arm.Parameters populated with parameter names and
